Trying to reduce bootloader size Moving signature functions in a separate file in a skycoin-api Removing object files from bootloader
