+++ hardware-wallet/skycoin-api/Makefile
@@ -28,6 +28,7 @@ INC+=-Itools/
 CFLAGS += -I. $(INC)
 
 SRCS += skycoin_crypto.c
+SRCS += skycoin_check_signature.c
 SRCS += $(shell ls tools/*.c)
 
 OBJS   = $(SRCS:.c=.o)
@@ -42,8 +43,8 @@ all: test_skycoin_crypto
 skycoin-crypto: skycoin_crypto.o $(OBJS)
 	$(CC) -rdynamic -shared $(CFLAGS) $(OBJS) -o libskycoin-crypto.so
 
-skycoin-crypto-link-static: skycoin_crypto.o $(OBJS)
-	$(CC) -rdynamic -shared -fPIC $(CFLAGS) $(OBJS) -o libskycoin-crypto.so
+libskycoin-crypto.a: skycoin_crypto.o $(OBJS)
+	$(AR) rcs libskycoin-crypto.a skycoin_crypto.o $(OBJS)
 
 test_skycoin_crypto: test_skycoin_crypto.o skycoin-crypto
 	$(CC) test_skycoin_crypto.o $(OBJS) -L. -lskycoin-crypto  $(TESTLIBS) -o test_skycoin_crypto
@@ -54,3 +55,5 @@ clean:
 	rm -f *.o test_skycoin_crypto
 	rm -f *.so
 	rm -f tools/*.o
+	rm -f trezor-tools/*.o
+	rm -f *.a
