+++ hardware-wallet/skycoin-api/skycoin_crypto.h
@@ -14,7 +14,6 @@ void generate_deterministic_key_pair(const uint8_t* seed, const size_t seed_leng
 void generate_base58_address_from_pubkey(const uint8_t* pubkey, char* address, size_t *size_address);
 void generate_bitcoin_address_from_pubkey(const uint8_t* pubkey, char* address, size_t *size_address);
 void generate_bitcoin_private_address_from_seckey(const uint8_t* pubkey, char* address, size_t *size_address);
-int recover_pubkey_from_signed_message(const char* message, const uint8_t* signature, uint8_t* pubkey);
 int ecdsa_skycoin_sign(const uint32_t nonce_value, const uint8_t *priv_key, const uint8_t *digest, uint8_t *sig);
 void tohex(char * str, const uint8_t* buffer, int bufferLength);
 #endif
