+++ hardware-wallet/tiny-firmware/bootloader/Makefile
@@ -10,21 +10,14 @@ OBJS += ../serialno.o
 OBJS += ../setup.o
 OBJS += ../util.o
 OBJS += ../memory.o
-OBJS += ../timer.o
 OBJS += ../gen/bitmaps.o
 OBJS += ../gen/fonts.o
 #end libtrezor.a
 
 #skycoin-crypto
-OBJS += $(TOP_DIR)vendor/skycoin-crypto/skycoin_crypto.o
-OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/base58.o
-OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/blake256.o
-OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/hasher.o
-OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/rand.o
-
+OBJS += $(TOP_DIR)vendor/skycoin-crypto/skycoin_check_signature.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/curves.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/hmac.o
-OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/ripemd160.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/bip32.o
 
 #end skycoin-crypto
